Code clean up (#65)
* Updated to ContactModel

* Updated to ProjectModel

* Updated TabType to PageModel

* Updated errors with constants

* Moved page components into own folder

* Updated TabPanel folder name

* Minor changes to clean up code
