--- src/constants/contact-type.ts
@@ -1,9 +1,0 @@
-export default class ContactTypeConstants {
-  static readonly EMAIL: string = 'Email';
-
-  static readonly GITHUB: string = 'GitHub';
-
-  static readonly LINKEDIN: string = 'LinkedIn';
-
-  static readonly RESUME: string = 'Resume';
-}
