+++ src/constants/contact.ts
@@ -1,0 +1,9 @@
+export default class ContactConstants {
+  static readonly EMAIL: string = 'Email';
+
+  static readonly GITHUB: string = 'GitHub';
+
+  static readonly LINKEDIN: string = 'LinkedIn';
+
+  static readonly RESUME: string = 'Resume';
+}
