--- src/types/component-helpers/tab.ts
@@ -1,10 +1,0 @@
-export default class Tab {
-  label: string;
-
-  component: JSX.Element;
-
-  constructor(label: string, component: JSX.Element) {
-    this.label = label;
-    this.component = component;
-  }
-}
