--- tests/components/contact/Contacts.test.tsx
@@ -1,27 +1,0 @@
-import { render } from '@testing-library/react';
-import Contacts from '../../../src/components/contact/Contacts';
-import ContactType from '../../../src/types/contact/contact';
-
-import { contacts } from '../../mocks/data-mock';
-
-describe('Contacts component', () => {
-  let getByTestId: any;
-
-  beforeEach(() => {
-    ({ getByTestId } = render(<Contacts contacts={contacts} />));
-  });
-
-  it('renders contact buttons with correct icons', () => {
-    expect(getByTestId('GitHubIcon')).toBeVisible();
-    expect(getByTestId('LinkedInIcon')).toBeVisible();
-    expect(getByTestId('PictureAsPdfIcon')).toBeVisible();
-  });
-
-  it('renders contact buttons with correct URLs', () => {
-    contacts.forEach((contact: ContactType, index: number) => {
-      const element = document.getElementById(`btn-${index}`) as HTMLButtonElement;
-
-      expect(element).toHaveAttribute('href', contact.url);
-    });
-  });
-});
