+++ tests/mocks/data-mock.js
@@ -1,8 +1,8 @@
-import ContactType from '../../src/types/contact/contact';
+import ContactModel from '../../src/models/contact/contact';
 
 export const contacts = [
-  new ContactType('GitHub', 'Contact description.', 'Contact value', 'github.com'),
-  new ContactType('LinkedIn', 'Contact description 2.', 'Contact value 2', 'linkedin.com'),
-  new ContactType('Resume', 'Contact description 3.', 'Contact value 3'),
-  new ContactType('None', 'Contact description 4.', 'Contact value 4'),
+  new ContactModel('GitHub', 'Contact description.', 'Contact value', 'github.com'),
+  new ContactModel('LinkedIn', 'Contact description 2.', 'Contact value 2', 'linkedin.com'),
+  new ContactModel('Resume', 'Contact description 3.', 'Contact value 3'),
+  new ContactModel('None', 'Contact description 4.', 'Contact value 4'),
 ];
